backend/requirements.txt
livekit-agents>=0.8.0
python-dotenv
openai>=1.0.0
